api/requirements.txt
# llama-index
llama-index-core==0.12.52
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.4.7
llama-index-readers-file==0.4.11

# api
fastapi==0.116.1

# server
uvicorn[standard]==0.35.0

# utils
python-dotenv==1.1.1
